(GH-3633, GH-3634) Renamed GroupBoxHelper and added a new attached property for header background
- Rename GroupBoxHelper to HeaderedControlHelper
- Added attached property HeaderBackground
