+++ src/MahApps.Metro/Styles/Clean/CleanGroupBox.xaml
@@ -6,7 +6,7 @@
         <Setter Property="BorderBrush" Value="{DynamicResource MahApps.Brushes.Gray7}" />
         <Setter Property="BorderThickness" Value="0.3" />
         <Setter Property="Controls:ControlsHelper.HeaderFontSize" Value="16" />
-        <Setter Property="Controls:GroupBoxHelper.HeaderForeground" Value="{DynamicResource MahApps.Brushes.Text}" />
+        <Setter Property="Controls:HeaderedControlHelper.HeaderForeground" Value="{DynamicResource MahApps.Brushes.Text}" />
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="GroupBox">
@@ -16,24 +16,26 @@
                             <RowDefinition Height="*" />
                         </Grid.RowDefinitions>
 
-                        <Controls:ContentControlEx Grid.Row="0"
-                                                   Margin="{TemplateBinding Padding}"
-                                                   Content="{TemplateBinding Header}"
-                                                   ContentCharacterCasing="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:ControlsHelper.ContentCharacterCasing)}"
-                                                   ContentStringFormat="{TemplateBinding HeaderStringFormat}"
-                                                   ContentTemplate="{TemplateBinding HeaderTemplate}"
-                                                   ContentTemplateSelector="{TemplateBinding HeaderTemplateSelector}"
-                                                   FontFamily="{TemplateBinding Controls:ControlsHelper.HeaderFontFamily}"
-                                                   FontSize="{TemplateBinding Controls:ControlsHelper.HeaderFontSize}"
-                                                   FontStretch="{TemplateBinding Controls:ControlsHelper.HeaderFontStretch}"
-                                                   FontWeight="{TemplateBinding Controls:ControlsHelper.HeaderFontWeight}"
-                                                   Foreground="{TemplateBinding Controls:GroupBoxHelper.HeaderForeground}"
-                                                   RecognizesAccessKey="True"
-                                                   SnapsToDevicePixels="{TemplateBinding SnapsToDevicePixels}" />
+                        <Grid Grid.Row="0" Background="{TemplateBinding Controls:HeaderedControlHelper.HeaderBackground}">
+                            <Controls:ContentControlEx Margin="{TemplateBinding Padding}"
+                                                       Content="{TemplateBinding Header}"
+                                                       ContentCharacterCasing="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:ControlsHelper.ContentCharacterCasing)}"
+                                                       ContentStringFormat="{TemplateBinding HeaderStringFormat}"
+                                                       ContentTemplate="{TemplateBinding HeaderTemplate}"
+                                                       ContentTemplateSelector="{TemplateBinding HeaderTemplateSelector}"
+                                                       FontFamily="{TemplateBinding Controls:ControlsHelper.HeaderFontFamily}"
+                                                       FontSize="{TemplateBinding Controls:ControlsHelper.HeaderFontSize}"
+                                                       FontStretch="{TemplateBinding Controls:ControlsHelper.HeaderFontStretch}"
+                                                       FontWeight="{TemplateBinding Controls:ControlsHelper.HeaderFontWeight}"
+                                                       Foreground="{TemplateBinding Controls:HeaderedControlHelper.HeaderForeground}"
+                                                       RecognizesAccessKey="True"
+                                                       SnapsToDevicePixels="{TemplateBinding SnapsToDevicePixels}" />
+                        </Grid>
 
                         <Grid Grid.Row="1"
                               HorizontalAlignment="Stretch"
-                              VerticalAlignment="Stretch">
+                              VerticalAlignment="Stretch"
+                              Background="{TemplateBinding Background}">
                             <Grid.ColumnDefinitions>
                                 <ColumnDefinition Width="Auto" />
                                 <ColumnDefinition Width="*" />
