+++ src/MahApps.Metro/Styles/Controls.GroupBox.xaml
@@ -9,12 +9,12 @@
     <Thickness x:Key="GroupBoxHeaderThemePadding">4</Thickness>
 
     <Style x:Key="MahApps.Styles.GroupBox" TargetType="{x:Type GroupBox}">
-        <Setter Property="Background" Value="{DynamicResource MahApps.Brushes.Accent}" />
         <Setter Property="BorderBrush" Value="{DynamicResource MahApps.Brushes.Accent}" />
         <Setter Property="BorderThickness" Value="1" />
         <Setter Property="Controls:ControlsHelper.ContentCharacterCasing" Value="Upper" />
         <Setter Property="Controls:ControlsHelper.HeaderFontSize" Value="{DynamicResource MahApps.Sizes.Font.Content}" />
-        <Setter Property="Controls:GroupBoxHelper.HeaderForeground" Value="{x:Null}" />
+        <Setter Property="Controls:HeaderedControlHelper.HeaderBackground" Value="{DynamicResource MahApps.Brushes.Accent}" />
+        <Setter Property="Controls:HeaderedControlHelper.HeaderForeground" Value="{x:Null}" />
         <Setter Property="Foreground" Value="{DynamicResource MahApps.Brushes.Black}" />
         <Setter Property="Padding" Value="4" />
         <Setter Property="SnapsToDevicePixels" Value="True" />
@@ -28,7 +28,7 @@
                         </Grid.RowDefinitions>
                         <Border x:Name="HeaderSite"
                                 Grid.Row="0"
-                                Background="{TemplateBinding Background}"
+                                Background="{TemplateBinding Controls:HeaderedControlHelper.HeaderBackground}"
                                 BorderBrush="{TemplateBinding BorderBrush}"
                                 BorderThickness="{TemplateBinding BorderThickness}"
                                 CornerRadius="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:ControlsHelper.CornerRadius), Converter={StaticResource CornerRadiusBindingConverter}, ConverterParameter={x:Static Converters:RadiusType.Bottom}}"
@@ -51,17 +51,17 @@
                                 <TextElement.Foreground>
                                     <MultiBinding Converter="{x:Static Converters:BackgroundToForegroundConverter.Instance}">
                                         <Binding Mode="OneWay"
-                                                 Path="Background"
+                                                 Path="(Controls:HeaderedControlHelper.HeaderBackground)"
                                                  RelativeSource="{RelativeSource TemplatedParent}" />
                                         <Binding Mode="OneWay"
-                                                 Path="(Controls:GroupBoxHelper.HeaderForeground)"
+                                                 Path="(Controls:HeaderedControlHelper.HeaderForeground)"
                                                  RelativeSource="{RelativeSource TemplatedParent}" />
                                     </MultiBinding>
                                 </TextElement.Foreground>
                             </Controls:ContentControlEx>
                         </Border>
                         <Border Grid.Row="1"
-                                Background="Transparent"
+                                Background="{TemplateBinding Background}"
                                 BorderBrush="{TemplateBinding BorderBrush}"
                                 BorderThickness="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=BorderThickness, Converter={StaticResource ThicknessBindingConverter}, ConverterParameter={x:Static Converters:ThicknessSideType.Top}}"
                                 CornerRadius="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:ControlsHelper.CornerRadius), Converter={StaticResource CornerRadiusBindingConverter}, ConverterParameter={x:Static Converters:RadiusType.Top}}"
