+++ src/MahApps.Metro/Styles/VS/Expander.xaml
@@ -333,7 +333,8 @@
         <Setter Property="Controls:ExpanderHelper.HeaderLeftStyle" Value="{StaticResource MahApps.Styles.ToggleButton.ExpanderHeader.Left.VisualStudio}" />
         <Setter Property="Controls:ExpanderHelper.HeaderRightStyle" Value="{StaticResource MahApps.Styles.ToggleButton.ExpanderHeader.Right.VisualStudio}" />
         <Setter Property="Controls:ExpanderHelper.HeaderUpStyle" Value="{StaticResource MahApps.Styles.ToggleButton.ExpanderHeader.Up.VisualStudio}" />
-        <Setter Property="Controls:GroupBoxHelper.HeaderForeground" Value="{DynamicResource MahApps.Brushes.Foreground}" />
+        <Setter Property="Controls:HeaderedControlHelper.HeaderBackground" Value="{x:Null}" />
+        <Setter Property="Controls:HeaderedControlHelper.HeaderForeground" Value="{DynamicResource MahApps.Brushes.Foreground}" />
         <Setter Property="Foreground" Value="{DynamicResource MahApps.Brushes.Foreground}" />
         <Setter Property="SnapsToDevicePixels" Value="True" />
         <Setter Property="Template">
@@ -347,6 +348,7 @@
                         <DockPanel x:Name="ExpanderRoot">
                             <ToggleButton x:Name="ToggleSite"
                                           Controls:ControlsHelper.ContentCharacterCasing="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:ControlsHelper.ContentCharacterCasing)}"
+                                          Background="{TemplateBinding Controls:HeaderedControlHelper.HeaderBackground}"
                                           BorderBrush="{TemplateBinding BorderBrush}"
                                           Content="{TemplateBinding Header}"
                                           ContentStringFormat="{TemplateBinding HeaderStringFormat}"
@@ -357,14 +359,14 @@
                                           FontSize="{TemplateBinding Controls:ControlsHelper.HeaderFontSize}"
                                           FontStretch="{TemplateBinding Controls:ControlsHelper.HeaderFontStretch}"
                                           FontWeight="{TemplateBinding Controls:ControlsHelper.HeaderFontWeight}"
-                                          Foreground="{TemplateBinding Controls:GroupBoxHelper.HeaderForeground}"
+                                          Foreground="{TemplateBinding Controls:HeaderedControlHelper.HeaderForeground}"
                                           IsChecked="{Binding IsExpanded, Mode=TwoWay, RelativeSource={RelativeSource TemplatedParent}}"
                                           SnapsToDevicePixels="{TemplateBinding SnapsToDevicePixels}"
                                           Style="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=(Controls:ExpanderHelper.HeaderDownStyle)}" />
                             <Border x:Name="ExpandSite"
                                     HorizontalAlignment="{TemplateBinding HorizontalContentAlignment}"
                                     VerticalAlignment="{TemplateBinding VerticalContentAlignment}"
-                                    Background="Transparent"
+                                    Background="{TemplateBinding Background}"
                                     BorderBrush="{TemplateBinding BorderBrush}"
                                     BorderThickness="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=BorderThickness, Converter={StaticResource ThicknessBindingConverter}, ConverterParameter={x:Static Converters:ThicknessSideType.Top}}"
                                     DockPanel.Dock="Bottom"
